Powershell script to do the task of terraform code
This is a powershell script executing AWS cli commands. In this scenario this script is capable of doing much more than the terraform code. The tf code is limited to creating replica key in one region, if we want to do it in lore than one we need to repeat this process. But here it is possible to create replica key in more than one region at once and can also be used where it is not possible to use terraform.